% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_postcode_lookup.R
\name{bulk_postcode_lookup}
\alias{bulk_postcode_lookup}
\title{Bulk postcode lookup}
\usage{
bulk_postcode_lookup(postcodes)
}
\arguments{
\item{postcodes}{Accepts a JSON object containing an array of postcodes. Accepts up to 100 postcodes.}
}
\description{
Returns a list of matching postcodes and respective available data.
}

