% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_reverse_geocoding.R
\name{bulk_reverse_geocoding}
\alias{bulk_reverse_geocoding}
\title{Reverse geocoding - NOT WORKING}
\usage{
bulk_reverse_geocoding(longitude, latitude)
}
\arguments{
\item{A}{list containing an array of geolocation objects.}
}
\value{
A list with available data.
}
\description{
Returns nearest postcodes for a given longitude and latitude. Accepts up to 100 geolocations.
}
\details{
This method requires a JSON object containing an array of geolocation objects to be POSTed.
Each geolocation object accepts an optional radius (meters) and limit argument.
Both default to 100m and 10 respectively. It also accepts a wideSearch argument.
}
\examples{
bulk_reverse_geocoding()
}

