% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_outcode_lonlat.R
\name{nearest_outcode_lonlat}
\alias{nearest_outcode_lonlat}
\title{Nearest outcodes given longitude and latitude}
\usage{
nearest_outcode_lonlat(longitude, latitude)
}
\arguments{
\item{longitude}{A string or numeric. Needs to have at least three decimal points.}

\item{latitude}{A string or numeric. Needs to have at least three decimal points.}
}
\value{
A list with available data.
}
\description{
Returns nearest outward codes for a given longitude and latitude.
The search is based on the relative distance of the outcode centroid.
}
\examples{
\donttest{
nearest_outcode_lonlat(0.127, 51.507)
}

}
\seealso{
\code{\link{postcode_lookup}} for documentation.
}
