% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_postcode.R
\name{nearest_postcode}
\alias{nearest_postcode}
\title{Nearest postcode}
\usage{
nearest_postcode(postcode)
}
\arguments{
\item{postcode}{A string.}
}
\value{
A list of geographic properties of the nearest postcode.
}
\description{
Returns nearest postcodes for a given postcode.
}
\details{
Optional Query Parameters

limit= (not required) Limits number of postcodes matches to return. Defaults to 10. Needs to be less than 100.

radius= (not required) Limits number of postcodes matches to return. Defaults to 100m. Needs to be less than 2,000m.
}
\examples{
nearest_postcode("EC1Y 8LX")

}
