% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_lookup.R
\name{place_lookup}
\alias{place_lookup}
\title{Place Lookup}
\usage{
place_lookup(code)
}
\arguments{
\item{code}{A string. Ordnance Survey National Grid code.}
}
\value{
A list with available places.
}
\description{
Lookup a place by code. Returns all available data if found. Returns 404 if place does not exist.
}
\examples{
place_lookup("osgb4000000074544700")

}
