% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_query.R
\name{place_query}
\alias{place_query}
\title{Place Query}
\usage{
place_query(place)
}
\arguments{
\item{place}{A string. Name of a place to search for.}
}
\value{
A list with available places.
}
\description{
Submit a place query and receive a complete list of places matches and associated data.
}
\examples{
place_query("Cambridge")

}
