% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_autocomplete.R
\name{postcode_autocomplete}
\alias{postcode_autocomplete}
\title{Postcode autocomplete}
\usage{
postcode_autocomplete(postcode)
}
\arguments{
\item{postcode}{A string.}
}
\value{
A list of suggested postcodes. Defaults to 10.
}
\description{
Convenience method to return an list of matching postcodes.
}
\details{
Optional Query Parameters

limit= (not required) Limits number of postcodes matches to return. Defaults to 10. Needs to be less than 100.
}
\examples{
postcode_autocomplete("E1")
}

