% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_lookup.R
\name{postcode_lookup}
\alias{postcode_lookup}
\title{Postcode lookup}
\usage{
postcode_lookup(postcode)
}
\arguments{
\item{postcode}{A string. Valid UK postcode.}
}
\value{
A data frame. Returns all available data if found. Returns 404 if postcode does not exist.
}
\description{
Lookup a postcode.
}
\examples{
postcode_lookup("EC1Y8LX")
postcode_lookup("EC1Y 8LX")

}
