% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_query.R
\name{postcode_query}
\alias{postcode_query}
\title{Postcode query}
\usage{
postcode_query(postcode)
}
\arguments{
\item{postcode}{A string.}
}
\value{
A list of geographic properties.
}
\description{
Submit a postcode query and receive a complete list of postcode matches and all associated postcode data.
}
\details{
Optional Query Parameters

query= (not required) aliases to q=

limit= (not required) Limits number of postcodes matches to return. Defaults to 10. Needs to be less than 100.
}
\examples{
postcode_query("EC1Y8LX")
}

