% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_validation.R
\name{postcode_validation}
\alias{postcode_validation}
\title{Postcode validation}
\usage{
postcode_validation(postcode)
}
\arguments{
\item{postcode}{A string.}
}
\value{
A list consisting of the HTTP response, and True or False (meaning valid or invalid respectively).
}
\description{
Convenience method to validate a postcode.
}
\examples{
postcode_validation("EC1Y 8LX")
}

