% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_place.R
\name{random_place}
\alias{random_place}
\title{Random Place}
\usage{
random_place()
}
\value{
A data frame describing a random place and all associated data.
}
\description{
Returns a random place and all associated data
}
\examples{
\donttest{
random_place()
}

}
\seealso{
\code{\link{place_lookup}} for documentation.
}
