% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_geocoding.R
\name{reverse_geocoding}
\alias{reverse_geocoding}
\title{Reverse geocoding}
\usage{
reverse_geocoding(longitude, latitude)
}
\arguments{
\item{longitude}{A string. Needs to have at least three decimal points.}

\item{latitude}{A string. Needs to have at least three decimal points.}
}
\value{
A list with available data.
}
\description{
Returns nearest postcodes for a given longitude and latitude.
}
\details{
Optional Query Parameters:
limit= (not required) Limits number of postcodes matches to return. Defaults to 10. Needs to be less than 100.
radius= (not required) Limits number of postcodes matches to return. Defaults to 100m. Needs to be less than 2,000m.
wideSearch= (not required) Search up to 20km radius, but subject to a maximum of 10 results. Since lookups over a wide area can be very expensive, we've created this method to allow you choose to make the trade off between search radius and number of results. Defaults to false. When enabled, radius and limits over 10 are ignored.
}
\examples{
reverse_geocoding("-3.158", "51.479")
}

