% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminated_postcode.R
\name{terminated_postcode}
\alias{terminated_postcode}
\title{Terminated postcode lookup}
\usage{
terminated_postcode(postcode = NULL)
}
\arguments{
\item{postcode}{A string. Terminated UK postcode.}
}
\value{
A data frame with data about terminated postcode.
}
\description{
Returns nearest postcodes for a given longitude and latitude.
}
\examples{
terminated_postcode("EC1Y 8LX") # exisitng postcode - returns 404
terminated_postcode("E1W 1UU") # terminated - from Wikipedia but returns 404

}
