% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_postcode_lookup.R
\name{bulk_postcode_lookup}
\alias{bulk_postcode_lookup}
\title{Bulk postcode lookup}
\usage{
bulk_postcode_lookup(...)
}
\arguments{
\item{...}{One or more vectors containing postcodes to look up. Up to 100 postcodes are accepted.
For only one postcode use \code{\link{postcode_lookup}}.}
}
\value{
A list
}
\description{
Returns a list of matching postcodes and respective available data.
}
\examples{
\donttest{
## Postcodes can be provided as individual arguments
bulk_postcode_lookup("PR30SG", "M456GN", "EX165BL")
## Or as one or more vectors:
bulk_postcode_lookup(c("PR30SG", "M456GN"), "EX165BL")
}
}
\seealso{
\code{\link{postcode_lookup}} for documentation.
}
