% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_reverse_geocoding.R
\name{bulk_reverse_geocoding}
\alias{bulk_reverse_geocoding}
\title{Bulk reverse geocoding}
\usage{
bulk_reverse_geocoding(geolocations)
}
\arguments{
\item{geolocations}{A list containing an array of objects to geolocate.
At least two elements needed.}
}
\value{
A list with the geocoded data.
}
\description{
Returns nearest postcodes for a given longitude and latitude. Accepts up to 100 geolocations.
}
\details{
This method requires a JSON object containing an array of geolocation objects to be POSTed.
Each geolocation object accepts an optional radius (meters) and limit argument.
Both default to 100m and 10 respectively. It also accepts a wideSearch argument.
}
\examples{

\donttest{
geolocations_list <- structure(
list(
geolocations = structure(
list(
longitude = c(-3.15807731271522, -1.12935802905177),
latitude = c(51.4799900627036, 50.7186356978817),
limit = c(NA, 100L),
radius = c(NA, 500L)),
.Names = c("longitude", "latitude", "limit", "radius"),
class = "data.frame",
row.names = 1:2)),
.Names = "geolocations")

bulk_reverse_geocoding(geolocations_list)
}

}
\seealso{
\code{\link{postcode_lookup}} for documentation.
}
