% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_outcode.R, R/nearest_postcode.R
\name{nearest_postcode}
\alias{nearest_postcode}
\title{Returns nearest outcodes for a given outcode.}
\usage{
nearest_postcode(postcode)

nearest_postcode(postcode)
}
\arguments{
\item{postcode}{A string.}

\item{postcode}{A string.}
}
\value{
A list of geographical properties.

A list of geographic properties of the nearest postcode.
}
\description{
Optional Query Parameters

Returns nearest postcodes for a given postcode.
}
\details{
limit= (not required) Limits number of postcodes matches to return. Defaults to 10. Needs to be less than 100.

radius= (not required) Limits number of postcodes matches to return. Defaults to 5,000m. Needs to be less than 25,000m.

Optional Query Parameters

limit= (not required) Limits number of postcodes matches to return. Defaults to 10. Needs to be less than 100.

radius= (not required) Limits number of postcodes matches to return. Defaults to 100m. Needs to be less than 2,000m.
}
\examples{
nearest_postcode("EC1Y 8LX")
nearest_postcode("EC1Y8LX")
nearest_postcode("EC1Y 8LX")
}

