% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outward_code_lookup.R
\name{outward_code_lookup}
\alias{outward_code_lookup}
\title{Outward code lookup}
\usage{
outward_code_lookup(outcode)
}
\arguments{
\item{outcode}{A string. The outward code representing the first half of any postcode (separated by a space).}
}
\value{
The list of geographical properties.
}
\description{
Geolocation data for the centroid of the outward code specified.
}
\examples{
\donttest{
outward_code_lookup("E1")
}

}
\seealso{
\code{\link{postcode_lookup}} for documentation.
}
