% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_autocomplete.R
\name{postcode_autocomplete}
\alias{postcode_autocomplete}
\title{Postcode autocomplete}
\usage{
postcode_autocomplete(postcode, limit = 10)
}
\arguments{
\item{postcode}{A string. Valid UK postcode.}

\item{limit}{An integer. Limits number of postcodes matches to return. Defaults to 10. Needs to be less than 100.}
}
\value{
A data frame with suggested postcodes.
}
\description{
Returns a data frame of matching postcodes.
}
\examples{
\donttest{
postcode_autocomplete("E1")
postcode_autocomplete("E1", limit = 11)
}

}
