% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_validation.R
\name{postcode_validation}
\alias{postcode_validation}
\title{Postcode validation}
\usage{
postcode_validation(postcode)
}
\arguments{
\item{postcode}{A string. Valid UK postcode.}
}
\value{
A logical vector: True or False (meaning respectively valid or invalid postcode).
}
\description{
Convenience method to validate a postcode.
}
\examples{
\donttest{
postcode_validation("EC1Y 8LX") # returns TRUE
postcode_validation("XYZ") # returns FALSE
}

}
