% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_postcode.R
\name{random_postcode}
\alias{random_postcode}
\title{Random postcode}
\usage{
random_postcode(outcode = NULL)
}
\arguments{
\item{outcode}{A string. Filters random postcodes by outcode. Returns null if invalid outcode. Optional.}
}
\value{
A random post code with corresponding characteristics.
}
\description{
Returns a random postcode and all available data for that postcode.
}
\examples{
random_postcode()
random_postcode("N1")

}
