% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminated_postcode.R
\name{terminated_postcode}
\alias{terminated_postcode}
\title{Terminated postcode lookup}
\usage{
terminated_postcode(postcode)
}
\arguments{
\item{postcode}{A string. Terminated UK postcode.}
}
\value{
A data frame with data about terminated postcode. NULL if postcode is active.
}
\description{
Returns month and year if a postcode was terminated or is no longer active.
}
\examples{
terminated_postcode("EC1Y 8LX") # exisitng postcode
terminated_postcode("E1W 1UU") # terminated

}
